function limit update
git-svn-id: file:///usr/local/svn/cpplib/trunk/cftal@2545 18305beb-b1e2-0310-bcd8-b14b5419b1d7

+++ include/cftal/math/func_constants.h
@@ -170,8 +170,8 @@ namespace cftal {
             tgamma_hi_inf() { return 3.504010009765625000000000e+01f; }
 
             // tgamma(x) == +- 0 for x<
-            static constexpr double
-            tgamma_lo_zero() { return -42.0; }
+            static constexpr float
+            tgamma_lo_zero() { return -42.0f; }
         };
 
         namespace impl {
